libvbisam/isbuild.c:isaddindex fixed iserrno should be EKEXISTS
when a key is tried to be added but it turns out to be a duplicate
of an already-existing key. This is the expected behavior from
the original C-ISAM.

The solution was to divide the keydesc validation in two stages: first
do basic keydesc integrity checks only (call ivbcheckkey with imode 1)
and then skip those checks and do a key duplicity check (use imode 2).

If the basic key check fails, we let the iserrno set by ivbcheckkey
stand, and go fail. If the duplicity check fails, we set iserrno to
EKEYXISTS and then go fail.