libvbisam/isopen.c: reactivated row counting for isindexinfo
RXW (Roger While) commented out counting rows, hard-coded di_nrecords
to return 0 records counted.

I have not been able to find a justification for this action. Only
possible issue may be the performance of tcountrows, which is probably
not relevant for small enough tables. Some Changelog entries point to
incorrect record counts returned, but that was for early 1.x versions
of the code.